get cpu frequency for all supported arch, the bad point is that we need to duplicate a portion of the code in opcontrol since we need always cpu frequency in opcontrol to get the default count.
Shell code tested on x86/alpha/sparc64 and through school test case on
other arch, C code tested on x86 and through school test case on other
arch

regards,
Phil

+++ gui/oprof_start.h
@@ -142,7 +142,7 @@ private:
 	op_cpu cpu_type;
 
 	/// CPU speed in MHz
-	unsigned long cpu_speed;
+	double cpu_speed;
 
 	/// total number of available HW counters
 	uint op_nr_counters;
